Create a metadata flextable for a dataset

Builds a formatted metadata table summarizing the structure and content of a dataset. The table includes variable names, types, units, descriptions, and summary statistics. It is designed to render well in HTML using the flextable package.

create_metadata_table(df, fontname = NULL, fontsize = NULL)

Arguments

df

A data.frame with optional metadata attributes.

fontname

Optional font name to apply to the entire table.

fontsize

Optional font size to apply to the entire table.

Value

A flextable object summarizing the dataset.

Details

The dataset can include metadata as attributes:

A "description" attribute on the dataset itself

"description" and "unit" attributes on individual columns

This function assumes that variable descriptions and units have been added as attributes using add_desc() and add_unit(), and that dataset-level descriptions have been set using base::attr().

See also

Examples

# Load required packages
library(flextable)
library(dplyr)
#> 
#> Attaching package: 'dplyr'
#> The following objects are masked from 'package:stats':
#> 
#>     filter, lag
#> The following objects are masked from 'package:base':
#> 
#>     intersect, setdiff, setequal, union

# Create dataset and define descriptions and units for variables
df1 <- iris %>%
  add_desc(Sepal.Length = "Sepal length [cm]",  # units in square brackets are stored as unit attributes
           Sepal.Width  = "Sepal width",        # unit can be added separately
           Petal.Length = "Petal.Length [cm]",
           Petal.Width  = "Petal.Width [cm]",
           Species      = "Four Iris species") %>%
  add_unit(Sepal.Width = "cm")  # add unit if not set by add_desc

# Add dataset-level description
attr(df1, "description") <- "This famous (Fisher's or Anderson's) iris data set gives the measurements in centimeters of the variables sepal length and width and petal length and width, respectively, for 50 flowers from each of 3 species of iris. The species are Iris setosa, versicolor, and virginica."

# Create metadata table
ft <- create_metadata_table(df1)
ft

Metadata for dataset df1

This famous (Fisher's or Anderson's) iris data set gives the measurements in centimeters of the variables sepal length and width and petal length and width, respectively, for 50 flowers from each of 3 species of iris. The species are Iris setosa, versicolor, and virginica. Summary shows: numeric = mean ± SD (min–max) with NA count; factor = levels; character = unique count. 150 observations of 5 variables

Variable

Type

Unit

Summary

Description

Sepal.Length

numeric

cm

5.84 ± 0.83 (4.30–7.90), NA: 0

Sepal length

Sepal.Width

numeric

cm

3.06 ± 0.44 (2.00–4.40), NA: 0

Sepal width

Petal.Length

numeric

cm

3.76 ± 1.77 (1.00–6.90), NA: 0

Petal.Length

Petal.Width

numeric

cm

1.20 ± 0.76 (0.10–2.50), NA: 0

Petal.Width

Species

factor

setosa, versicolor, virginica

Four Iris species

# Save table as HTML
flextable::save_as_html(ft, path = "df1_metadata.html")